Multiple bug fixes and improvements
"createFolders" to "theFolders" due to function and variable being the same name causing conflict.

"directory" to "labelFolder" as this was confusing naming scheme. Referenced "celebrityFolder" for this naming scheme

Updated the "nullDirectories" function as needed for the newer variables as well as with better default variables.

Updated main variable section with correct variables; moved "labelFolder" to the top, since I'm now referencing that for the extra folders

Added additional Python arguments